+++ test/igraphs.jl
@@ -1,6 +1,6 @@
 @testset "igraphs" begin
     @test IGraphAlgorithm() isa AbstractGraphAlgorithm
-    
+
     # Test error hint for sir_model
     try
         sir_model(SimpleGraph(3), 0.1, 0.1, 10)
@@ -8,7 +8,7 @@
     catch e
         @test e isa MethodError
         @test e.f === sir_model
-        
+
         msg = sprint(showerror, e)
         @test contains(msg, "This function requires the IGraphs.jl package to be loaded")
     end