+++ CHANGELOG.md
@@ -4,6 +4,7 @@ We follow SemVer as most of the Julia ecosystem. Below you might see the "breaki
 
 ## unreleased
 - `is_articulation(g, v)` for checking whether a single vertex is an articulation point
+- Integration with the `igraph` C library via `IGraphs.jl` for high-performance algorithm implementations
 
 
 ## v1.14.0 - 2026-02-26